+++ R/discomb.R
@@ -820,7 +820,7 @@ discomb <- function(TE, seTE,
                         p0$treat1[o], p0$treat2[o], level.comb,
                         X.matrix, C.matrix, B.matrix,
                         Q, df.Q.additive, df.Q.diff,
-                        n, sep.trts)
+                        n, sep.trts, sep.comps)
   
   
   ##
@@ -847,7 +847,7 @@ discomb <- function(TE, seTE,
                         p1$treat1[o], p1$treat2[o], level.comb,
                         X.matrix, C.matrix, B.matrix,
                         Q, df.Q.additive, df.Q.diff,
-                        n, sep.trts)
+                        n, sep.trts, sep.comps)
   
 
   NAs <- rep(NA, length(res.f$comparisons$TE))
@@ -973,6 +973,22 @@ discomb <- function(TE, seTE,
               upper.Comb.random = unname(res.r$combinations$upper),
               statistic.Comb.random = unname(res.r$combinations$statistic),
               pval.Comb.random = unname(res.r$combinations$p),
+              ##
+              Comb.fixed.2comps = res.r$combinations.2comps$TE,
+              seComb.fixed.2comps = res.f$combinations.2comps$seTE,
+              lower.Comb.fixed.2comps = res.f$combinations.2comps$lower,
+              upper.Comb.fixed.2comps = res.f$combinations.2comps$upper,
+              statistic.Comb.fixed.2comps =
+                res.f$combinations.2comps$statistic,
+              pval.Comb.fixed.2comps = res.f$combinations.2comps$p,
+              ##
+              Comb.random.2comps = res.r$combinations.2comps$TE,
+              seComb.random.2comps = res.r$combinations.2comps$seTE,
+              lower.Comb.random.2comps = res.r$combinations.2comps$lower,
+              upper.Comb.random.2comps = res.r$combinations.2comps$upper,
+              statistic.Comb.random.2comps =
+                res.r$combinations.2comps$statistic,
+              pval.Comb.random.2comps = res.r$combinations.2comps$p,
               ##
               Q.additive = Q.additive, 
               df.Q.additive = df.Q.additive,