+++ R/nma.additive.R
@@ -3,7 +3,7 @@ nma.additive <- function(TE, weights, studlab,
                          level.comb,
                          X, C.matrix, B.matrix,
                          Q, df.Q.additive, df.Q.diff,
-                         n, sep.trts) {
+                         n, sep.trts, sep.comps) {
   
   
   m <- length(TE)
@@ -39,16 +39,37 @@ nma.additive <- function(TE, weights, studlab,
   se.theta <- sqrt(diag(C.matrix %*% Lplus %*% t(C.matrix)))
   names(theta) <- names(se.theta)
   ##
+  ## Calculate estimates for all potential combinations of two
+  ## components (theta.2comps)
+  ##
+  n.beta <- length(beta)
+  C.2comps <- abs(createB(ncol = n.beta))
+  colnames(C.2comps) <- names(beta)
+  theta.2comps <- as.vector(C.2comps %*% beta)
+  se.theta.2comps <- sqrt(diag(C.2comps %*% Lplus %*% t(C.2comps)))
+  ##
+  nam.beta <- names(beta)
+  k <- 0
+  names.2comps <- vector(mode = "numeric", length = choose(n.beta, 2))
+  ##
+  for (i in 1:(n.beta - 1)) {
+    for (j in (i + 1):n.beta) {
+      k <- k + 1
+      names.2comps[k] <- paste(nam.beta[i], nam.beta[j], sep = sep.comps)
+    }
+  }
+  names(theta.2comps) <- names(se.theta.2comps) <- names.2comps
+  ##
   ## delta = treatment estimates for observed comparisons
   ##
   delta <- as.vector(X %*% beta) # = B.matrix %*% theta = H %*% TE
   se.delta <- unname(sqrt(diag(X %*% Lplus %*% t(X))))
   ##
-  ## delta.all = all direct and indirect treatment estimates
+  ## delta.all(.matrix) = all direct and indirect treatment estimates
   ##
   B.full <- createB(ncol = n)
-  X.full <- B.full %*% C.matrix
-  colnames(X.full) <- colnames(C.matrix)
+  X.all <- B.full %*% C.matrix
+  colnames(X.all) <- colnames(C.matrix)
   ##
   labels <- colnames(B.matrix)
   ##
@@ -62,37 +83,42 @@ nma.additive <- function(TE, weights, studlab,
     }
   }
   ##
-  rownames(X.full) <- lab
+  rownames(X.all) <- lab
   ##
-  delta.full <- as.vector(X.full %*% beta)
-  se.delta.full <- sqrt(diag(X.full %*% Lplus %*% t(X.full)))
-  names(delta.full) <- names(se.delta.full)
+  delta.all <- as.vector(X.all %*% beta)
+  se.delta.all <- sqrt(diag(X.all %*% Lplus %*% t(X.all)))
+  names(delta.all) <- names(se.delta.all)
   ##
-  delta.all <- se.delta.all <- matrix(0, ncol = n, nrow = n)
+  delta.all.matrix <- se.delta.all.matrix <- matrix(0, ncol = n, nrow = n)
   ##
   k <- 0
   ##
   for (i in 1:(n - 1)) {
     for (j in (i + 1):n) {
       k <- k + 1
-      delta.all[i, j] <-  delta.full[k]
-      delta.all[j, i] <- -delta.full[k]
-      se.delta.all[i, j] <- se.delta.all[j, i] <- se.delta.full[k]
+      delta.all.matrix[i, j] <-  delta.all[k]
+      delta.all.matrix[j, i] <- -delta.all[k]
+      se.delta.all.matrix[i, j] <-
+        se.delta.all.matrix[j, i] <- se.delta.all[k]
     }
   }
   ##
-  colnames(delta.all) <- rownames(delta.all) <-
-    colnames(se.delta.all) <- rownames(se.delta.all) <- labels
+  colnames(delta.all.matrix) <- rownames(delta.all.matrix) <-
+    colnames(se.delta.all.matrix) <- rownames(se.delta.all.matrix) <- labels
   
   
   comparisons <- c(list(studlab = studlab, treat1 = treat1, treat2 = treat2),
                    ci(delta, se.delta, level = level.comb))
   ##
-  all.comparisons <- ci(delta.all, se.delta.all, level = level.comb)
+  all.comparisons <- ci(delta.all.matrix, se.delta.all.matrix,
+                        level = level.comb)
   ##
   components <- meta::ci(beta, se.beta, level = level.comb)
   ##
   combinations <- meta::ci(theta, se.theta, level = level.comb)
+  ##
+  combinations.2comps <-
+    meta::ci(theta.2comps, se.theta.2comps, level = level.comb)
   ##
   ## Test of total heterogeneity / inconsistency:
   ##
@@ -145,6 +171,7 @@ nma.additive <- function(TE, weights, studlab,
               all.comparisons = all.comparisons,
               components = components,
               combinations = combinations,
+              combinations.2comps = combinations.2comps,
               ##
               Q.additive = Q.additive,
               df.Q.additive = df.Q.additive,